# Conflicts:
#	data/src/main/java/com/yapp/growth/data/api/GrowthApi.kt
#	data/src/main/java/com/yapp/growth/data/di/DataSourceModule.kt
#	data/src/main/java/com/yapp/growth/data/di/RepositoryModule.kt
#	data/src/main/java/com/yapp/growth/data/internal/response/FixedPlanResponseImpl.kt
#	data/src/main/java/com/yapp/growth/data/mapper/FixedPlanMapper.kt
#	data/src/main/java/com/yapp/growth/data/response/FixedPlanResponse.kt
#	presentation/src/main/java/com/yapp/growth/presentation/ui/main/manage/ManageScreen.kt

+++ domain/src/main/java/com/yapp/growth/domain/entity/Plan.kt
@@ -1,7 +1,7 @@
 package com.yapp.growth.domain.entity
 
 sealed class Plan(
-    open val id: Int,
+    open val id: Long,
     open val title: String,
     open val isLeader: Boolean,
     open val category: String, // type?
@@ -9,7 +9,7 @@ sealed class Plan(
     open val place: String,
 ) {
     data class WaitingPlan(
-        override val id: Int,
+        override val id: Long,
         override val title: String,
         override val isLeader: Boolean = false,
         override val category: String,
@@ -28,7 +28,7 @@ sealed class Plan(
     )
 
     data class FixedPlan(
-        override val id: Int,
+        override val id: Long,
         override val title: String,
         override val isLeader: Boolean = false,
         override val category: String,
